# listen_stream_server/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Listen Stream API
  version: 1.0.0
  description: |
    Listen Stream 后端全路由契约规范。
    - 所有时间字段均为 ISO 8601 UTC 字符串（`2026-02-20T12:00:00Z`）
    - 所有列表接口统一分页格式：`{ data, total, page, size }`
    - Proxy 路由的 `data` 字段内容透传自第三方 API，枚举值不在本规范中固化
    - 错误响应统一格式：`{ code: string, message?: string }`
servers:
  - url: http://localhost:8001
    description: auth-svc (开发)
  - url: http://localhost:8002
    description: proxy-svc (开发)
  - url: http://localhost:8003
    description: sync-svc (开发)
  - url: http://localhost:8004
    description: admin-svc (开发)

# ─────────────────────────────────────────────
# Security
# ─────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT，Claims 中须含 `aud` 字段。
        - 用户接口：`aud = ["user"]`，由 auth-svc 签发，使用 USER_JWT_SECRET 验证。
        - 管理员接口：`aud = ["admin"]`，由 admin-svc 签发，使用 ADMIN_JWT_SECRET 验证。

  # ─────────────────────────────────────────────
  # Reusable schemas
  # ─────────────────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      required: [code]
      properties:
        code:
          type: string
          description: 机器可读错误码
        message:
          type: string
          description: 可选的人类可读描述

    # ── Pagination ──
    PaginationMeta:
      type: object
      required: [total, page, size]
      properties:
        total:
          type: integer
          minimum: 0
        page:
          type: integer
          minimum: 1
        size:
          type: integer
          minimum: 1
          maximum: 100

    # ── Auth ──
...
    Device:
      type: object
      required: [id, deviceId, platform, lastActiveAt, createdAt]
      properties:
        id:
          type: string
        deviceId:
          type: string
        platform:
          type: string
          enum: [android, ios, desktop, tv]
        lastActiveAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    # ── Sync ──
...
    AdminUserRole:
      type: string
      enum: [USER, ADMIN, SUPER_ADMIN]
    AdminUserItem:
      type: object
      required: [id, phone, role, disabled, deviceCount, createdAt]
      properties:
        id:
          type: string
        phone:
          type: string
        role:
          $ref: '#/components/schemas/AdminUserRole'
        disabled:
          type: boolean
        deviceCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    AdminUserListResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationMeta'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/AdminUserItem'

    # ── Admin: Config ──
...
paths:
...
  /user/devices:
    get:
      tags: [User - Device]
      summary: 当前用户的设备列表
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 设备列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /user/devices/{deviceId}:
    delete:
      tags: [User - Device]
      summary: 用户主动下线指定设备
      security:
        - BearerAuth: []
      parameters:
        - name: deviceId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: 已下线
        '403':
          description: 不能操作不属于自己的设备
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: DEVICE_NOT_OWNED
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ══════════════════════════════════════════
  # Admin — admin-svc :8004
  # ══════════════════════════════════════════
...
  /admin/users:
    get:
      tags: [Admin - Users]
      summary: 用户列表（分页+手机号搜索）
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1, minimum: 1 }
        - name: size
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
        - name: phone
          in: query
          schema: { type: string }
          description: 手机号前缀搜索（LIKE）
      responses:
        '200':
          description: 用户列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/users/{id}/role:
    put:
      tags: [Admin - Users]
      summary: 修改用户角色（仅 super_admin）
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  $ref: '#/components/schemas/AdminUserRole'
      responses:
        '200':
          description: 已更新
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/users/{id}/status:
    put:
      tags: [Admin - Users]
      summary: 启用/禁用用户（禁用立即吊销所有 RT）
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [disabled]
              properties:
                disabled:
                  type: boolean
      responses:
        '200':
          description: 已更新
        '403':
          description: 不能禁用自己
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: CANNOT_DISABLE_SELF
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/devices/{deviceId}:
    delete:
      tags: [Admin - Users]
      summary: 管理员吊销指定设备（DEL RT + 发送 device.kicked）
      security:
        - BearerAuth: []
      parameters:
        - name: deviceId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: 已吊销
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...